Moved around the packages.
Package folders were too deep and annoying. Fixed that.
