+++ nbproject/project.properties
@@ -69,7 +69,7 @@ jnlp.signed=false
 jnlp.signing=
 jnlp.signing.alias=
 jnlp.signing.keystore=
-main.class=wright.state.scheduler.WrightStateScheduler
+main.class=WSS.WrightStateScheduler
 # Optional override of default Application-Library-Allowable-Codebase attribute identifying the locations where your signed RIA is expected to be found.
 manifest.custom.application.library.allowable.codebase=
 # Optional override of default Caller-Allowable-Codebase attribute identifying the domains from which JavaScript code can make calls to your RIA without security prompts.
